clean up extraneous fragments in app & add anim cancel button
