+++ CustomGlyphNotifications/app/src/main/java/com/ubercab/MainActivity.java
@@ -1,46 +1,28 @@
 package com.ubercab;
 
-import android.app.Notification;
-import android.app.NotificationChannel;
-import android.app.NotificationManager;
 import android.content.Context;
 import android.os.Bundle;
-
-import com.google.android.material.snackbar.Snackbar;
-
 import androidx.appcompat.app.AppCompatActivity;
-
-import android.util.Log;
-import android.view.View;
-
-import androidx.core.app.NotificationCompat;
-import androidx.core.app.NotificationManagerCompat;
 import androidx.navigation.NavController;
 import androidx.navigation.Navigation;
 import androidx.navigation.ui.AppBarConfiguration;
 import androidx.navigation.ui.NavigationUI;
-
 import com.ubercab.databinding.ActivityMainBinding;
-
 import android.view.Menu;
 import android.view.MenuItem;
 
 public class MainActivity extends AppCompatActivity {
 
-    private final String CHANNEL_ID = "TestChannel";
-
     private AppBarConfiguration appBarConfiguration;
     private ActivityMainBinding binding;
 
-    private int minsLeft = 20;
-
-    Context curContext;
+    public static Context context;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
 
-        curContext = this;
+        context = this;
 
         binding = ActivityMainBinding.inflate(getLayoutInflater());
         setContentView(binding.getRoot());
@@ -50,45 +32,6 @@ public class MainActivity extends AppCompatActivity {
         NavController navController = Navigation.findNavController(this, R.id.nav_host_fragment_content_main);
         appBarConfiguration = new AppBarConfiguration.Builder(navController.getGraph()).build();
         NavigationUI.setupActionBarWithNavController(this, navController, appBarConfiguration);
-
-        CreateNotificationChannel();
-
-        binding.fab.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                Log.d("hahah", "Sending notification");
-
-                Bundle extras = new Bundle();
-                extras.putString("android.title", "Pickup in " + minsLeft + " mins");
-
-                // Build notification
-                Notification.Builder notifBuilder = new Notification.Builder(curContext, CHANNEL_ID)
-                        .setSmallIcon(R.drawable.ic_launcher_foreground)
-                        .setContentText("haha")
-                        .addExtras(extras)
-                        .setAutoCancel(true);
-
-                // Send notification
-                NotificationManagerCompat notificationManager = NotificationManagerCompat.from(curContext);
-                notificationManager.notify(24, notifBuilder.build());
-
-                minsLeft--;
-            }
-        });
-    }
-
-
-    private void CreateNotificationChannel() {
-            CharSequence name = "Hello chan";
-            String description = "Hello chan";
-            int importance = NotificationManager.IMPORTANCE_DEFAULT;
-            NotificationChannel channel = new NotificationChannel(CHANNEL_ID, name, importance);
-            channel.setDescription(description);
-            channel.setImportance(NotificationManager.IMPORTANCE_DEFAULT);
-            // Register the channel with the system; you can't change the importance
-            // or other notification behaviors after this
-            NotificationManager notificationManager = getSystemService(NotificationManager.class);
-            notificationManager.createNotificationChannel(channel);
     }
