+++ CustomGlyphNotifications/app/src/main/res/navigation/nav_graph.xml
@@ -7,22 +7,7 @@
 
     <fragment
         android:id="@+id/FirstFragment"
-        android:name="com.ubercab.FirstFragment"
+        android:name="com.ubercab.GlyphProgressFragment"
         android:label="@string/first_fragment_label"
-        tools:layout="@layout/fragment_first">
-
-        <action
-            android:id="@+id/action_FirstFragment_to_SecondFragment"
-            app:destination="@id/SecondFragment" />
-    </fragment>
-    <fragment
-        android:id="@+id/SecondFragment"
-        android:name="com.ubercab.SecondFragment"
-        android:label="@string/second_fragment_label"
-        tools:layout="@layout/fragment_second">
-
-        <action
-            android:id="@+id/action_SecondFragment_to_FirstFragment"
-            app:destination="@id/FirstFragment" />
-    </fragment>
+        tools:layout="@layout/fragment_first"/>
 </navigation>
